deprecate `StringUtil.blockStringValue` in `sangria-core`
It's believed that this method is unused outside of the parser.

Moves the parser's copy of the implementation from the `Tokens` trait
into a new `Lexical` object, from which it can be easily made visible if
needed in the future.  Moves the tests over from `core` into `parser`.

Also fixes some IDE warnings.
