+++ modules/core/src/test/scala/sangria/util/StringUtilSpec.scala
@@ -55,119 +55,4 @@ class StringUtilSpec extends AnyWordSpec with Matchers with StringMatchers {
       suggestionList("descritpion", options) should be(Seq("description"))
     }
   }
-
-  "blockStringValue" should {
-    "removes uniform indentation from a string" in {
-      blockStringValue(
-        """
-          |    Hello,
-          |      World!
-          |
-          |    Yours,
-          |      GraphQL.
-          |""".stripMargin
-      ) should equal(
-        """Hello,
-          |  World!
-          |
-          |Yours,
-          |  GraphQL.""".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-    }
-
-    "removes empty leading and trailing lines" in {
-      blockStringValue(
-        """
-          |
-          |    Hello,
-          |      World!
-          |
-          |    Yours,
-          |      GraphQL.
-          |
-          |""".stripMargin
-      ) should equal(
-        """Hello,
-          |  World!
-          |
-          |Yours,
-          |  GraphQL.""".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-    }
-
-    "removes blank leading and trailing lines" in {
-      val spaces = " " * 10
-
-      blockStringValue(
-        s"""
-           |$spaces
-           |    Hello,
-           |      World!
-           |
-           |    Yours,
-           |      GraphQL.
-           |$spaces
-           |""".stripMargin
-      ) should equal(
-        """Hello,
-          |  World!
-          |
-          |Yours,
-          |  GraphQL.""".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-    }
-
-    "retains indentation from first line" in {
-      blockStringValue(
-        """    Hello,
-           |      World!
-           |
-           |    Yours,
-           |      GraphQL.
-           |""".stripMargin
-      ) should equal(
-        """    Hello,
-          |  World!
-          |
-          |Yours,
-          |  GraphQL.""".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-    }
-
-    "does not alter trailing spaces" in {
-      val spaces = " " * 10
-
-      blockStringValue(
-        s"""
-           |    Hello,$spaces
-           |      World!$spaces
-           |    $spaces
-           |    Yours,$spaces
-           |      GraphQL.$spaces
-           |""".stripMargin
-      ) should equal(
-        s"""Hello,$spaces
-           |  World!$spaces
-           |$spaces
-           |Yours,$spaces
-           |  GraphQL.$spaces""".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-    }
-
-    "handles empty strings" in {
-      val spaces = " " * 10
-
-      blockStringValue(
-        s"""
-           |$spaces
-           |
-           |$spaces
-           |""".stripMargin
-      ) should equal(
-        "".stripMargin
-      )(after.being(strippedOfCarriageReturns))
-
-      blockStringValue("") should equal("")
-    }
-  }
 }
